\name{design}
\alias{design}
\docType{data}
\title{Experimental design for the example RRBS dataset}

\description{
  The RRBS dataset is from 16 samples with two factors (case and cell),
  each has two levels (so it's a 2x2 design). 
}
\usage{data(RRBS)}
\examples{
data(RRBS)
RRBS
design
}
\keyword{datasets}
