% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dss.R
\name{dss}
\alias{dss}
\title{Digital Skill Score - DSS}
\usage{
dss(x, p = NULL, group_by = "month")
}
\arguments{
\item{x:}{Number of transactions.}

\item{p:}{Digital proportion.}

\item{group_by:}{Must group_by 'day', 'month' (Default), 'quarter' or 'year'?}
}
\value{
Digital Skill Score, `dss = log(x+1)*(1-log(p))`
}
\description{
Calculates the Digital Skill Score for a customer with `x` transactions with digital proportion `p`.
}
\examples{
library(dss)

dss(dss_dataset2)
}
