\name{RRBS}
\alias{RRBS}
\docType{data}
\title{An example dataset for multiple factor design}
\description{
  The dataset contains RRBS data for 5000 CpG sites from 16 samples. The
  experimental design is provided in the 'design' data frame.
}
\usage{data("RRBS")}
\examples{
data(RRBS)
RRBS
design
}
\keyword{datasets}
