\name{seqData}
\alias{seqData}
\docType{data}
\title{
  A simulated 'SeqCountData' object.
}
\description{
  The object is created based on simulation for 1000 genes and two
  treatment groups with 4 replicates in each group.
}
\usage{data(seqData)}

\examples{
data(seqData)
seqData
}
\keyword{datasets}
