% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dss.R
\name{dss}
\alias{dss}
\title{Digital Skill Score - DSS}
\usage{
dss(x, p = NULL, group_by = "month", keep_tables = FALSE, round_dec = 4)
}
\arguments{
\item{x:}{Number of transactions.}

\item{p:}{Digital proportion.}

\item{group_by:}{Must group_by 'day', 'month' (Default), 'quarter' or 'year'?}
}
\value{
Digital Skill Score, `dss = log(x+1)*(1-log(p))`.
}
\description{
Calculates the Digital Skill Score for a customer with `x` transactions with digital proportion `p`.
}
\examples{
library(dss)

# If x is a vector, must enter p
dss(1)
dss(1, 0.9)
dss(1, 0.9, round_dec = Inf)
dss(c(1,3,21,8), c(.9,.3,.1,.5))

# If x is a data frame
dss(dss_dataset1, group_by = 'day')
dss(dss_dataset1, group_by = 'month')
dss(dss_dataset1, group_by = 'month', keep_tables = TRUE)
dss(dss_dataset1, group_by = 'year')
dss(dss_dataset1, group_by = c('year', 'quarter'))
}
