% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dss_dataset_doc.R
\docType{data}
\name{dss_dataset1}
\alias{dss_dataset1}
\title{DSS Dataset 1}
\format{
A data frame with 15 rows and 4 variables:
\describe{
  \item{id}{Customer ID.}
  \item{time}{Timestamp.}
  \item{transaction}{Transaction type.}
  \item{digital_proportion}{Transaction digital proportion.}
}
}
\source{
Filipe J. Zabala.
}
\usage{
library(dss)
data(dss_dataset1)
}
\description{
Syntetic dataset containing the expected structure to apply on `dss::dss()`.
}
\keyword{datasets}
